feature: add 'id' attribute to the book
This was one to facilitate the removal of a book
from the library. What this id does is auto-increment itself
each time a book is added.

This id will also be assigned to the button that is
used to delete a button.  When the button is pressed,
the id of the button is taken to find it in the array.