feat: remove 'ID' row from library display
The function displayLibrary() was modified so that it does not show the
row that indicates the id of the book. This was done due to personal preference.

What was done to achieve this is via DOM manipulation, change the style attribute
'visibility' to 'hidden'.  Another thing that was done to achieve this was
to not append to the body of the table the cell originally created to display the id of the book.