fix: fix displayLibrary() to create dynamically the table
The table, and the table container were finally created dynamically.

When debugging, the div with the id 'table-container' was being created again
every time a book was added to the library.  A solution for this was to
validate in the 'displayLibrary' function if the table-container existed.
If it does not exist, then it is created, and appended to the body of
the HTML file.  If it does exist, then the 'clearTable' function was called

Another thing that stood out was that the row containing the table headers was being
duplicated every time a book was added to the library (along with the table being duplicated as well).
A solution for this was to create the header of the table outside of the loop.