feat: hide 'id' row on the table display
The row displaying the id of each book in the table was
hidden due to personal preference.  This was done by setting the
'visibility' style attribute of the variables 'bookIDHeaderCell', and
'bookIDCell' variables.

Also, when removing a book from the library, the book was not being
removed, and the length of the array was not changing.
What was happening was that the element was now an empty string.
To solve this, the length of the array was reduced after removing
book in the 'removeBookFromLibrary' function